requirements.txt
aiofiles==0.3.2
aiohttp==3.2.1
async-timeout==3.0.0
attrs==18.1.0
chardet==3.0.4
colorama==0.3.9
httptools==0.0.11
idna==2.6
idna-ssl==1.0.1
more-itertools==4.1.0
multidict==4.3.1
Pillow==5.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
pyzbar==0.1.7
sanic==0.7.0
six==1.11.0
websockets==4.0.1
yarl==1.2.4
